Combined taxonomy getters
